// christmas-task/src/modules/enum.ts
export enum Sort {
  ascending = 'По возрастанию года выпуска',
  descending = 'По убыванию года выпуска',
  all = 'Все игрушки',
  fromAZ = 'По имени А-Я',
  fromZA = 'По имени Я-А'
}

export enum Shape {
  bell = 'колокольчик',
  ball = 'шар',
  cone = 'шишка',
  snowflake = 'снежинка',
  figurine = 'фигурка'
}

export enum Color {
  white = 'белый',
  yellow = 'желтый',
  red = 'красный',
  blue = 'синий',
  green = 'зелёный'
}